backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pandas>=2.2.0
numpy>=1.26.0
ollama>=0.3.0
python-dotenv>=1.0.0
scikit-learn>=1.5.0
plotly>=5.24.0
rapidfuzz>=3.0.0
pytest>=7.0.0
# optional: openai>=1.0.0  # for OpenAI fallback when OPENAI_API_KEY is set
